# be/src/config/swagger.yaml
openapi: 3.0.3
info:
  title: E-Commerce API
  version: 1.0.0
  description: |
    E-Commerce backend — Node.js · Express 5 · PostgreSQL · Knex · VNPay · GHN.

    **Auth:**
    - Customer endpoints use `Bearer <accessToken>` (user JWT, secret = `JWT_SECRET`)
    - Staff endpoints use `Bearer <staffToken>` (staff JWT, secret = `STAFF_JWT_SECRET`)
    - IPN and webhook endpoints are unsigned (verified internally by HMAC / GHN token)
servers:
  - url: http://localhost:3000/api
    description: Local development
...
components:
  securitySchemes:
    userAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Customer JWT from /auth/login or /auth/register
    staffAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Staff JWT from /admin/staff/auth/login
  schemas:
...
    Paginated:
      type: object
      properties:
        data: { type: array, items: {} }
        meta:
          type: object
          properties:
            total: { type: integer }
            page: { type: integer }
            limit: { type: integer }
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, default: 20 } }
        - { name: sort, in: query, schema: { type: string } }
        - { name: category_id, in: query, schema: { type: integer } }
        - { name: minPrice, in: query, schema: { type: number } }
        - { name: maxPrice, in: query, schema: { type: number } }
        - { name: attribute_id, in: query, schema: { type: integer } }
      responses:
        "200":
          {
            description: OK,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Paginated" } },
              },
          }
    post:
      tags: [Products]
      summary: Create product (Staff)
      security: [{ staffAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, description]
              properties:
                name: { type: string, maxLength: 20 }
                description: { type: string, maxLength: 100 }
                category_ids: { type: array, items: { type: integer } }
      responses:
        "201":
          {
            description: Created,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: object,
                        properties:
                          {
                            data:
                              {
                                type: object,
                                properties: { product_id: { type: integer } },
                              },
                          },
                      },
                  },
              },
          }
...
  /products/{id}:
    get:
      tags: [Products]
      summary: Product detail with variants, images, price
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: OK }
        "404": { description: Not found }
    put:
      tags: [Products]
      summary: Full update (Staff)
      security: [{ staffAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 20 }
                description: { type: string, maxLength: 100 }
                category_ids: { type: array, items: { type: integer } }
      responses:
        "200": { description: Updated }
    patch:
      tags: [Products]
      summary: Partial update (Staff)
      security: [{ staffAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 20 }
                description: { type: string, maxLength: 100 }
      responses:
        "200": { description: Updated }
    delete:
      tags: [Products]
      summary: Delete product (Staff)
      security: [{ staffAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204": { description: Deleted }
...